Fixed the issue that Architecture.Attributes was always empty: Made Attribute a subclass of Class. Added method to identify attributes in TypeFactory to add them to the Architecture as Attribute and not just as Class. Added GetAttributeOfType() extension method to Architecture. Added Test for Attributes property of Architecture.
Changed the Equality operator of Class to just check equality of Type such that references at Classes which are not in the library are still equal. That was not the case before because Classes outside the library are instantiated with every property (IsAbstract,IsSealed,...) set to false by default, even though they might actually be true. Another solution was most likely not possible because we can't find out these properties if typeReference.Resolve() does return null in TypeFactory, Line 55.
Since it is impossible to have two classes with the same full name but different properties the change should be fine.

Signed-off-by: Fritz Brandhuber <fritz.brandhuber@tngtech.com>