Change rule evaluation to also require positive results by default
The change is intended to make it easier to detect ineffective rules,
as you can see in the example unit test. This change obviously breaks
some tests, so it can be defeated per rule with a call to

    .WithoutRequiringPositiveResults().

Should().Exist() and NotExist() rules automatically defeat the check
so their error messages should remain stable. However, the logic to check
for (non)exitence had to be modified which may cause inconvenience.

The implemetation basically treats an empty result set as a non-passing
result (i.e. a violation).

The change breaks a lot of unit tests, mostly those which iterate
over a bunch of test types and are tautologies. This issue is adressed
in a subsequent commit.

Signed-off-by: Simon Thum <sthum@meddv.de>